Modify setInterval and setTimeout eval blocked test for Cobalt.
This modifies the tests for unsafe-eval blocked when passing to
setInterval and setTimeout to use separate script tags for the
blocked usage with evaluated DOMString versus the not blocked usage
with a handler function (without eval) that should pass.

This allows these tests to be enabled for Cobalt, which does not
implement the DOMString overrides for setInterval() and setTimeout(),
which results in interruption of the script when they are
encountered.

Change-Id: I943edc8a7785097c9145cec3618c601f1b34710f

Former-commit-id: 166f4fd7ae94cfb4135080375f430fa142e14e52