Allow multiple RenderTargets to be used with the Skia HardwareRasterizer.
Previously, if you changed RenderTargets skia may or may not render
correctly depending upon whether the original RenderTarget used was of the
same size as the new RenderTarget because it used the same SkSurface
throughout the life of rasterizer. Now, a map is used to map RenderTargets
to SkSurfaces so that each RenderTarget can have it's own surface to draw
into. The map is an LRU cache of size 2 so that we don't hold hold onto a
bunch of SkSurfaces for too long if they're used once and then dropped;
this size could be increased arbitrarily if there becomes a need.

Change-Id: I4d803d2a4c3c9e306a8d89cb1313feec962f7d33
(cherry picked from commit fe3ac14d4e2e2bf315aea661e1bf175a4a1edebc)

Former-commit-id: d140173d12ee1446f71ef954e8df3fa0902b7631