Add clang and gcc targets for x64x11 variant.
This adds various clang (3.3, 3.6) and gcc (4.2, 4.4) toolchains for
the x64x11 variant.

For clang 3.3 and gcc 4.2, the compiler is built from source. For the
other compiler versions, apt-get can be used for installation.

The compiler flags are separated from the other variables in
gyp_configuration.gypi, so that different values can be used for
different compiler versions.

For the other linux builds, the linux/shared configuration now
explicitly points to clang/clang++ in third_party/llvm-build. It should
no longer be necessary to that path to the PATH environment variable.

Change-Id: I5063ecae22f2b66afbc37b765cc7a063dca2daf7

Former-commit-id: 6718db9d2ec94bb2913ae27ca988d1fd1fb192b1