Use JavaScript instead of CSS to show splash screen spinner.
Originally, CSS Animations were used to have the spinner appear
after 2 seconds.  This change uses JavaScript instead.  As it was
before, Cobalt didn't know that the animation that would reveal
the spinner was not active most of the time, and so it would
re-render the splash screen each frame since an animation is
playing.

Also change the 2 second delay before spinner appearance to
3 seconds, to make it possible for Littlefield to load with
no spinner in an ideal network situation.

Improves startup time on Littlefield by ~0.5 seconds.

Change-Id: I1e5bf5254f97cbb4381a9a6f61f34ba94dbdd633

Former-commit-id: 590a0dbc2a4d3eb11e5ea4e28260d86cd37f364b