Adds modes -memory_tracker=runtime and -memory_tracker=startup
When mode is runtime then the memory_tracker is sampled once a second
and this is dumped to stdout.

When mode is startup then the memory_tracker is sampled frequently
and then CSV stats of the engine are dumped to stdout.

If no mode is specified then a helpful message instructing the user
the available modes and what they do are printed and then the engine
requests to exit it's process.

Change-Id: I3412d648d644c4ee6d47922d4b8f1e98e94f7490

Former-commit-id: f9a836f8b43b1bf34318af3743868e40fd02f5c5