Adds modes -memory_tracker=runtime and -memory_tracker=startup
When mode is runtime then the memory_tracker is sampled once a second
and this is dumped to stdout.

When mode is startup then the memory_tracker is sampled frequently
and then CSV stats of the engine are dumped to stdout.

If no mode is specified then a helpful message instructing the user
the available modes and what they do are printed and then the engine
requests to exit it's process.

Change-Id: I3412d648d644c4ee6d47922d4b8f1e98e94f7490

Former-commit-id: 744304bdb5a5a2fb2e65e89d3fa5a808ff1b81e2