Make PunchThroughVideoNode render deterministically.
As it was, the results of rendering a punch out node depended on
whether a player was active or not, meaning that it may punch out
or it may render nothing.  This meant that we could not assume that
the same render tree would rasterize to the same output because it
also depends on whether a player is active or not.  This caused
artifacts when rendering only dirty rectangles, like for example
having the video only appear within the spinner for a time.

https://jira.portal.roku.com:8443/browse/YT-105

Change-Id: I38e5e4b45b39cad9cfb0d277f7f4b1b8060c23fd

Former-commit-id: 952bf857745e8763b57c0d41990a0ab9975b1c87