Fix a couple issues in SkFontMgr_cobalt
Previously, matchFaceNameOnlyIfFound() was not threadsafe as a result
of mathPostScriptName() not being threadsafe. The new function,
matchFaceName() handles the locking itself and is fully threadsafe.

Additionally, the font face name maps held on to pointers to
|SkFontStyleSetEntry_Cobalt| objects. However, these objects can be
destroyed by their owning |SkFontStyleSet_Cobalt| objects if they
fail to successfully load. As a result of this, there'd be stale
pointers in the maps that could potentially cause Cobalt to crash
later on. This has been changed so that only a pointer to the
|SkFontStyleSet_Cobalt| object is retained, and the entries are walked
to find the correct entry when matchFaceName() is called.

Also, did some minor cleanup of the code.

Change-Id: Ie937288ef9a930dbf73d60ee5fd0c3e5a2e5a2e3

Former-commit-id: dc57b12c913815952b4c23de23aa2869655e0227