PS3: Use localized strings in application
b/33002854

Adds a LocalizedStrings object to the PS3 application and uses it to
localize the subtitle for the save data.

Also replaces a usage of a C++11 function not supported by PS3 and fixes
a bug in convert_i18n_data.py that was converting the '\n' into "\r\n"
on the Windows host.

Change-Id: Iedc0be2c8ee4411654d34b826c33a99c56e85a85

Former-commit-id: 5c50b15f38be51d2e9f3534c995822acdf20717b