Fix renderer glitch where it appears to snap back in time.
This was actually intentional, though on non-really-slow devices
it did not appear to be a problem.

Basically, if layout submits a new render tree that has a
timestamp of being in the past from the renderer's perspective,
the renderer will snap back to that render tree submission's time.
This CL makes it not do that.  The logic was originally added
to deal with the case where no animations are playing and the user
hits a button: we might as well snap to the start of the new
submission to show the complete animation. Unfortunately that
logic doesn't play nice when a new submission arrives while an
animation is in progress: the snap back will make the animation
look like it glitches back in time.

With this change, we never "snap back in time" anymore, but we may
miss some of an animation.

Fixes b/33454125.

Change-Id: Ibf5f98cd0b3e1eab53d8b7907fc0f4e379d76352

Former-commit-id: 1c575176f217f5a7e056b85e7b0fa4a46cc330fd