Adds compressed time series tool for memory tracker.
This allows the developer to inspect the growth of memory regions for the
lifetime of the engine.

A compressed histogram is a type of logorithmic scale histogram. Data
to the right of the midpoint is at full resolution samples, but the data to
the left of the midpoint becomes progressively lower resolution. This allows
high fidelity reporting (before an out of memory exception) but still allows
the developer to see the memory trend.

Change-Id: I394c8b07567b1d9635f6a93da98a5eb2aa4bcf77

Former-commit-id: 048bcc1a9f58cbaf1b6fbf8747f0c3e7beb339e0