Handle fetching completion on the loader thread
Before this CL, if a fetch was started and finished, it would post a
completion callback on to the thread that started the fetch.  This was
non-ideal in some cases where the original calling thread was busy, and
could not process the data that was downloaded from the fetch.

Addresses b/33565232

Change-Id: I4ec8967e6309caa0929f14717ddcf26b70fe0a16

Former-commit-id: ea870831edc1a64d421bae46e72e7ee3390a4aea