sync_to_build_id.py to sync to a Cobalt build ID
Often times, it's useful to sync one's tree to a build before
a build broke.

With gclient, this is made harder by the fact that there's
multiple repos.

However, long ago someone made a mechanism to post the
current set of gclient hashes to the airlock appengine
machine so as to tag them with a build number.

This tool does the reverse: it sets your client to
be configured as the specified build number was configued.

This tool will fail if any of the git repos are dirty,
and will tell you which branches where changed and where
they were before they were changed.

Change-Id: Ice92b4b5e718e5852ef1ec596fc2ed4caae2883e

Former-commit-id: a0865b78895bbde11216c6723f1edbfaa1a64f70