Adds MemoryTrackerDebugCallback, which allows filtering allocations.
This greatly aids debugging of memory allocations. The intended use case
is to be able to filter allocations by memory scope and size so that
allocations can be checked.

Unit tests have been added to ensure that this works as expected.

Note: the git review system is showing that comments have been removed.
I'm not sure why this is, but probably has something to do with moving
the cl to this repo via git pull and then removing the comments.

Change-Id: I56e06dc8183b8f5387ff9e2c320bf8b80aae52b2

Former-commit-id: 7a431115752b1d2807590e4002fcd6ecf4520877