Fix raspi bug with char being unsigned
With ARM, chars are unsigned by default. This causes issues throughout
the scanner, which makes assumptions that the chars it operates on will
be signed.

For more information on this, see:
http://blog.cdleary.com/2012/11/arm-chars-are-unsigned-by-default/

To fix this, the gcc build option 'fsigned-char' has been added to
raspi1/gyp_configuration.gypi, so that the raspi builds always treat
chars as signed.

Change-Id: Idd0a715d15d34b07e00d5c8abdc18ce9ce29c809

Former-commit-id: 163a34ab445939c0e470c51ec8478188ef2faa1e