Linux dev_input: use all available keyboards
b/33646114

In the interest of seamlessly supporting bluetooth
keyboards on raspi, open all keyboard devices in
/dev/input and accept events from all of them.

Previously, the code would search through /dev/input/by-id
for the first symlink containing "event-kbd" that it found
and would open that. However, manually pairing a bluetooth
keyboard with "bluetoothctl" does not cause such a symlink
to be created.

So instead, simply look through all the top-level devices
in /dev/input and accept input from all that support
sending a keyboard space event.

Change-Id: I22dd3dd9d0e86c82e1c5d25ec1691e0a5cd7956a

Former-commit-id: e62a476e28bc5fbe4d8036f8fcc2895fcc69cefb