Android: fix SbMemoryMapTest
SbMemoryMapTest.CanExecuteMappedMemoryWithExecFlag was failing,
not for some weird SELinux policy, but simply because on a 32-bit
system we could end up with a negative number for our intptr_t.
Then when we added SB_MEMORY_PAGE_SIZE to it, we ended up
with an end before the beginning!

Just use uintptr_t here instead. We don't want signed arithmetic
here.

Change-Id: I0ad83a17880632cc81aa17e4fb88d0086529262e

Former-commit-id: 91271573c72776c1bccdd209b67518ccfaa2154a