Improve event tracking
The event tracking has been modified to work properly with the
webdriver benchmarks.

Previously, if a second event occured while the the first event was
in flight, the first event would end. However, in the case where
the first event triggered a layout which had not yet occurred, the
layout would incorrectly be attributed to the second event. This is
an issue with the webdriver benchmarks, because mouse down and mouse
up events occur in quick succession. While the mouse down event
triggers the layout, the layout was incorrectly being attributed to
the mouse up event. Now, if an event is currently being tracked, the
second event is ignored by the tracker.

A second issue with the previous logic is that it tracked the inject
event duration within window. However, in the case where the event is
directly being injected into a specific element, it did not go through
window and no injection timing was done. This is what occurs with the
webdriver benchmarks, meaning that JS times were not being tracked in
the benchmarks. This has now been modified so that the timing occurs
within the WebModule logic itself, rather than in the objects it calls.

A final issue with the previous logic is that events that did not
trigger layouts would not stop being tracked when the event finished
being injected. This meant that if a render tree was later produced
for another reason, it would incorrectly be ascribed to the event.
Now, if an event is injected and does not cause a new render tree to
be pending, tracking on it immediately ends. Whether or not an event
caused a new render tree to be produced is also recorded in a new
event cval, allowing metrics to easily differentiate events that
trigger new render trees with those that do not.

Change-Id: I2e32c8c3b31d0da262544478f70c6aa254cce73b

Former-commit-id: 395fb5fac4791e71e86da9f93c2033f53d249058