Android: Flesh out the Starboard application and window.
It almost runs, but crashes in layout when it can't get ICU. Fixing that
is too much scope creep for this change, so it's next.

Use the framework NativeActivity and the NDK android_native_app_glue
that spawns a thread for us, and simplifies the bookkeeping that we have
to do.

Minor refactor of the base Starboard Application:
1. Allow postponing sending the Starboard Start event until the Android
   app is fully running and has a Window.
2. Allow the app to dispatch the events so that it can indicate to keep
   keep pumping events until completely out of the Android lifecycle.

Use the none/stub type for GL and rasterizer for now since we don't
actually have a working GL window yet.

Move the Starboard Android application into the correct namespace.

Add an 'all_src' symlink so you can get to the rest of the code outside
the 'apk' directory in Android Studio.

Change-Id: I48c071104e3380e80e1300f2e676ee75a0c16c82

Former-commit-id: 1bf3b4a5d47772ec8ac3e5dcad81f71056eac7aa