Disable resource cache callbacks during shutdown
It is possible for resource cache loading callbacks to occur during
the shutdown process. This can happen on DOM objects even after the
window has been destroyed as a result of the JS engine holding on to
references to DOM objects. These callbacks then attempt to access a
stale Document object, resulting in a crash.

To fix this, resource cache callbacks are now disabled early in the
shutdown process.

I believe that this resolves b/33776258, b/33804740, and b/33771293.

Change-Id: I425dfb5c2801fe9a9fadefbf23b17c3110b3e392

Former-commit-id: 95f9d2ab28a86c9b784d75ce5da6de5806826835