Handle fetching completion on the loader thread
Before this CL, if a fetch was started and finished, it would post a
completion callback on to the thread that started the fetch.  This was
non-ideal in some cases where the original calling thread was busy, and
could not process the data that was downloaded from the fetch.

Addresses b/33565232
Original CL: go/lbreview/59495

The original CL was reverted due to a latent race condition.  This CL
fixes that race condition, and this change has passed over 100 hours of testing
under |input fuzzer|.

Change-Id: I8a91115ff987efa1cdb4369c3f459e4f7ea9c363

Former-commit-id: 685ce6b468776619ced2b5c78137a868d9556bb9