Change bool CVal return values to "1" and "0"
Previously the bool CVal was returning "true" and "false". However,
this made it so that special handling was required for them in
the webdriver tests and would be required by them in ***REMOVED***, since
the return value wouldn't automatically convert to true and false.
Simply returning "1" and "0" allows cleaner usage.

Change-Id: I14188f80d7c6dcb4b12ba7e3052d7f93145e749c

Former-commit-id: dd27c207d68f5a1a99cb9f9c943e680101302ca5