Add tracking of delay from event to video playing
A new cval has been added for tracking the amount of time taken
between the start of an event and when the video starts playing,
which indicates the browse to watch time and will be used in
webdriver tests. To generate this value, HtmlMediaElement now
calls into DomStatTracker when it is able to start playing its
video.

Additionally, did a bit of cleanup to the stat trackers.

Change-Id: I23aebed284ab98605adfee670a68547083857a52

Former-commit-id: 8e3d56227fec53bf2d80d9832d19dca11b1ffd8e