Remove incorrect check
The number of child element is 2, and not 1.  This test exists to
make sure that we do not crash if a charset is specified, and we do not
need to test the child element length.

Previously, the parser was returning 1.  However, after
go/lbreview/59825, the parser is returning 2 items (which is correct).

Change-Id: Ib2eaf8642a654673abbc0665329aca4789e72725

Former-commit-id: 63b4e8056c3da6a72dbb214a72caf9c6da18ea33