Add CValCollectionEntryStats and CValCollectionTimerStats
CValCollectionEntryStats allows entries to be added to a collection.
When either the max size is reached or Flush() is explicitly called,
CVal stats are generated for the current collection and the
collection's entries are cleared. Stats tracked are count, average,
min, max, standard deviation, 25th, 50th, 75th, and 95th percentiles.

CValCollectionTimerStats is a wrapper for CValCollectionEntryStats. It
provides Start() and Stop() calls which add TimeDelta entries to its
internal CValCollectionStats object. It also provides a Flush() call
for explicitly updating the stats and clearing the collection.

Additionally, unit tests were added for CValCollectionEntryStats,
CValCollectionTimerStats, CValTimeIntervalEntryStats, and
CValTimeIntervalTimerStats.

Change-Id: I5656d869d306afb721e25687a905563a392f63cf

Former-commit-id: 319e873d2e7bcc99fa2f8057e6a4f753a25351d0