Add timing of active animations
Added a new collection timer to Pipeline to track animations from when
they start until they expire. This will be used for tracking fps
during animations.

Additionally, the rasterization timer has been changed from an interval
timer to a collection timer. The reason for this is that the collection
timer can track more interesting info (because it stores the entries
until the flush), and works better on platforms that do not require
submissions even when the render tree is unchanged.

Change-Id: Id49bf7a4509af1fbf0180ca1a361eeb479ab037c

Former-commit-id: 787a6e352473d06feb316bf6ef8cf9ee3a0d5984