Android: Rework how lifecycle commands change Starboard app state.
Allows the app to be re-launched when the framework has not killed the
previous process. It resumes the UI wherever it was left last time,
which generally isn't the right UX in cases where it may have been
suspended for hours/days.

The Activity must run as "singleTask" since the Starboard application is
a singleton. Otherwise it won't work right (i.e. crash) if launched by
different means (e.g. launcher vs. apps preference panel).

Suspend/resume the Starboard app based on the window lifecycle. Then,
only when we do have a window, sync the Starboard app state to the
Activity lifecycle.

Make a SbWindowPrivate structure to hold the ANativeWindow pointer
rather than identity mapping the type to SbWindow. The ANativeWindow
will change when suspending and resuming but the SbWindow must remain
the same.

Info log all commands we receive from the native app glue.

Change-Id: I2061e9adc476c26e8e20d4c235a3b97868008c3d

Former-commit-id: 5ba8e0cb75160f1b2363b914c97aa7fedf0df5cd