Move clang to unified location.
The clang toolchain is moved from third_party/ to the unified
toolchain location.  The version string is appended to allow
collision-free building of branches that use different compiler
versions.  The version string is now specified in gyp_utils.py
instead of update.sh

This adds GetHostCompilerEnvironment() that returns the environment
variables for the host compiler, and handles the clang toolchain and
goma prerequisites.  Goma usage is switched from masquerade mode to
'gomacc' prepend mode, allowing goma to work even when the used clang
toolchain is not (or not correctly) in the path.  This also removes
the need to compare clang version numbers.

Replace CheckClangVersion() with EnsureClangAvailable() that uses
update.sh to ensure the given clang version is available.

This moves duplicated initialization of asan, tsan, and mtm variables
into a single location in GetVariables() in starboard.py

And 'git cl format' the files.

b/34166454

Change-Id: Ib5e8e580045837f2aabd9fc1b9e6239253848994

Former-commit-id: ed0307650e4d6ca91f6292a323b11c2f8a1eae50