Remove windows-1252-html converter mapping
In HTML5, windows-1252-html converter is used when ASCII converter is
requested.  This is because some webpages where mistakenly marking
windows-1252 text as latin-1.  Usually this was due to smart quotes in
windows-1252 being inserted by word processing software.  Since, we
intend to use our browser with sites that correctly set the encoding on
their text, we can avoid using the data based converter, and use
simpler algorithmic converter.  This can be done by defining
UCONFIG_NO_LEGACY_CONVERSION in ICU's gyp file.

Without this CL (and the data generated by ICU after this CL), ASCII conversion
in net_unittests fails and a test fails when UCONFIG_NO_LEGACY_CONVERSION is
defined in ICU.gyp.  With this CL (and the data files generated by this CL),
the test passes.

Change-Id: I020c3bb222478cf51de0cdc4b763cc04a81e9645

Former-commit-id: 8889e500fefb148c7c3ff36ca34b48aa7c04311b