Add flag to disable using goma.
This adds a flag that can disable trying to use goma, for compilers
on platforms where gomacc is available, but goma doesn't support the
specific compiler.

Note that since goma recognizes compilers using hashes on the
binaries, that means that any compiler toolchain built locally will
not be supported by goma.

b/34359233

Change-Id: Ib1971503913fb50106496a8f119835140635dffb

Former-commit-id: 2d3ff53032e7912ab46e930eee74a647437db924