Remove requirement that webdriver port end line
Previously, the regular expression searching for the webdriver port
required the port number to end the line. However, the PS3 is appending
an additional character at the end of the lines, causing this to break.
The end of the line requirement is being removed from the regular
expression so that it can handle this case.

Change-Id: Ied00806aa5a341d2ecd89248dfc67ea191bdb85d

Former-commit-id: 4d67105182dc8f4edff6aafdf5e32e0d34d0c832