Re-order OnRenderTreeProduced() calls
Previously, OnRenderTreeProduced() was being called on
WebModuleStatTracker after render tree produced callbacks were called.
However, these callbacks do a large amount of work when the debug
console is enabled and very little work when it is not, resulting in
an increased discrepancy between the times that QA would report and
the times that Gold would actually experience. The
WebModuleStatTracker calls has been moved before the callbacks so that
the tracked QA times will closer match Gold.

Change-Id: Ib3e91dc4739302cd6f6346d867ae52297af6c7cc

Former-commit-id: 25d83dcb9e50e29498c0031f34e82b4824a3752b