Add configuration for Raspi-2
This CL adds new raspi-2 configuration target, which produces more
tailored code for raspi 2.  Some changes are enabling code generation
for the armv7 processor, and changing |configuration_public.h| to
specify 4 cores.  Neon support was disabled in this build, since it
showed some degradations in preliminary benchmarks.  This can be easily
enabled afterwards after benchmarks show that it is a beneficial change.

Change-Id: I7543b44b968e134a4472ad76dad28ac8f8bd157a

Former-commit-id: 09cc382757c72fc86bc55ec1dbea093c8487544a