Android: sdk_utils.py auto-downloads SDK
sdk_utils.py (nee ndk_utils.py) now knows how to
download/install/update Android SDK automatically.

It will ensure that the tools and platform files that
it needs are available and use the "android update sdk" tool
if they are not.

The SDK will be installed in $COBALT_TOOLCHAINS/AndroidSdk and the
NDK will now be installed in $COBALT_TOOLCHAINS/AndroidSdk/ndk-bundle

This should suffice to enable building of the whole Android .apk package
on buildbot.

Change-Id: I39b9a59b3f3603f1e7400193d9157ffcc01b2a7c

Former-commit-id: 4cac42349df99bc14fe8d738c64ee35eb40d15bc