Add webdriver benchmarks for a variety of events
Webdriver benchmarks have been added for browse to watch, browse to
search, browse to guide, browse horizontal navigations, browse
vertical navigations, watch to browse, search to browse, and guide to
browse.

All of the cases use event recorders to record their event stats,
guaranteeing uniformity across tests in what is recorded. However,
only tests with animations record animation stats, and only browse
to watch records video stats. Also, to avoid noise in the data,
events that have font file loads are not included in the results.

With all of the stats being tracked, the mean, 25th, 50th, 75th, and
95th percentiles are recorded.

Additionally, added the ability to record the mean of a result array
and added unit tests for the mean functionality.

Change-Id: Ic397cbec2741614380b30717ef0af6e3c38e11d2

Former-commit-id: 502d3f3bba8cbaeed5cce46ecca2c6729e2c32b8