Create a linux-x64x11-lib platform for building Cobalt into a shared library.
This can be linked into the Youtube VR engine for running Cobalt in the
YouTube VR app. This is is just a starting shell at this point for
creating something that can be built, linked into a ***REMOVED***-built
cc_binary and run.

The intention is for most of the actual exported logic to be shared across the
platforms we would care about integrating with the YouTube VR engine in the
forseeable future (Windows, Xbox 2/Scorpio (?), Linux), but for now
it is kept entirely within Linux until we have another platform
directory to add to.

Change-Id: I2cddbf8c08d9106b5dac42f89695dbe1ffc8f434

Former-commit-id: 68c6e9ac16e53b69acb18e6278617041b18ac844