Move media2 from / to /cobalt.
The files were in MEDIA_REBASE_55 branch in the root folder.  Now they
are moved to /cobalt to become part of the cobalt media stack.
They are copied as is without any modification to serve as the base of
further modification and cleanup on including path, including guard,
and namespace etc.

b/34275215

Change-Id: I2b3685a76ec252de1480e58c3c18138fdb956ebb

Former-commit-id: 24bb6b2da68cfe7d929fc1c9c43c099fea2cbb1c