Change font config parsing to be done in memory
Previously, the parsing of the config file in
SkFontConfigParser_cobalt.cc was handled by retrieving data from
the file line by line and running xmlParseChunk() on each chunk of
data. However, given that the file will never be larger than a couple
dozen kb, this has been changed to read the entire file into memory
and then parse all of it in a single pass with xmlSAXUserParseMemory().
This significantly speeds up config parsing.

This resolves b/34752236.

Change-Id: Id5f1456b99d0d13d7d5bd04f0bc16cdadbcdcc03

Former-commit-id: c9bf997043ca622ddeca8476fd00a85832e7dc78