Fix performance bug in rule matching
Previously, during rule matching in ForEachChildOnNodes(), the
attributes of the element were retrieved at the start of the function.
However, this requires the creation of a NamedNodeMap for the element
and the attributes are almost never needed. This has been changed so
that the attributes are only requested when an attribute selector
exists.

There was also an additional bug in this code with the refptr to
NamedNodeMap being stored by reference in ForEachChildNodes() after it
was retrieved. However, the element itself holds a weak ptr to the
NamedNodeMap that it creates, meaning that the NamedNodeMap would be
destroyed as it was being returned and the reference that
ForEachChildOnNodes() was holding on to was actually stale. This has
been changed so that the refptr is now stored by value, guaranteeing
that the NamedNodeMap will remain in memory.

On the PS3, this change reduced the UpdateComputedStyle() duration
during vertical navigations from 57ms down to 37ms.

Change-Id: I483d4d54b6178098a2e196fd79790a622f31ffe6

Former-commit-id: 0847ed925f8b229ca7f840f7b1f261c65f465e0d