Android: Bump up to Starboard API 3
We'll want the SbSpeechSynthesis* stuff for WAI-ARIA, which
is in API version 3.

Additionally, we need to have symbols for the SbImageDecode* stuff,
which we're not currently planning on supporting.

This CL causes Android to link against stub versions of SbImageDecode*
and SbSpeechSynthesis*. The next CL will have concrete implementations
of SbSpeechSynthesis* but we'll continue with the stub of
SbImageDecode*.

Change-Id: I92aff012cdc8c1d376aa56e1bc31fe13e5935b36

Former-commit-id: 50d77b0e940bbdc67f48652c0db77081cb590363