Starboard: remove SB_HAS_SYMBOLIC_LINKS
b/34164642

Basically: the current implementation of SbFileGetPathInfo and SbFileDelete
check for symlinks if SB_HAS_SYMBOLIC_LINKS is enabled, but this API
ends up being insufficient.

For example, Android places it's application
directory inside something with symlinks. A common unit test base walks
up the directory hierarchy calling SbFileGetPathInfo(). When the caller
sees a symlink where it wants to see a directory, it fails.

If we had an application that really cared about symlinks, we'd need
to expose distinct functions for stat and lstat, instead of switching.

But everything in our application today just wants to treat symlinks as
a the real file they link to.

This fixes a lot of Android unit tests.

Change-Id: I701927dbb09dda418dce65cf09ce24c69e652eb9

Former-commit-id: 1e8bb8eb13d857ec0267b7b50b0e4928e8d2bf31