Optimize webdriver benchmarks
Previously, the webdriver benchmarks retrieved an extremely large
number of CVals, each of which requires a script to be run via
Selenium, which substantially slowed it down on certain platforms,
such as the PS3 and PS4.

This has been changed so that the event cvals are now collected
together within a single dictionary string CVal, and the
rasterize animation entries are collected together within a single
list string CVal. All rasterize animation entries being in a single
CVal has the added benefit of giving the webdriver benchmarks access
to all of the entries, which allows it to simplify its recorder logic,
since all value recorders can be treated the same now.

On the PS3, this reduces the benchmark time from 20 minutes down to
13 minutes. On the PS4, this reduces the time from 17 minutes down to
10 minutes.

Change-Id: Id03f78ac3b29b951b398fb4cefba1d501dffd027

Former-commit-id: 4df8cf3758587268b67fca3f8040fc10ecf84976