Make order of tests run by "all.py" deterministic
Previously, the tests within the same directory as "all.py" would be
run in an arbitrary order. This causes issues because "time_to_shelf"
only produces meaningful results if it is run first, and any bugs
related to the order of the tests would inconsistently appear. This
has been changed so that the tests are explicitly added in "all.py",
and run in the added order.

Change-Id: Ia8ed6d60de749b9d75d6b035a2d8b16ce1a9d0cd

Former-commit-id: ac6e52830d73dc3d102576bd66881ec19364f1b4