Remove unnecessary CVal queries from time_to_shelf
time_to_shelf was previously using the Cobalt.Lifetime CVal to
determine the elapsed time for loading a URL. However, there are
Python calls available that will do this, so there was no reason
to rely on Cobalt; additionally, the Cobalt queries added noise,
since they have to wait for an additional Selenium round trip. The
queries have been replaced with Python calls.

Change-Id: I0febd5181c6f5cf24f5dc6d7ccb789072166f056

Former-commit-id: 369e07cfe809bbdc9f26c26a97443c2e3353a4ad