Revert "Add log entry for when a new WindowDriver is created"
Unfortunately, this change broke the benchmarks on Linux and the PS4. It
seems that a sleep is still needed on some platforms even after the
window is created. I'm going to revert this and address the issue in a
different way.

However, the change to time_to_shelf.py has been retained.

This reverts commit 136e0e4401ef86a161454216e5bd9bafdaae770b.

Change-Id: I3ad8a0f5e8ff83ed00f1b0ba82f6d59a2bb865bb

Former-commit-id: 6167122892e2d8e84499975241cccb90a9fd739e