Modify tv_testcase_runner to launch Cobalt with url
Previously, the test runner would launch Cobalt without a URL
parameter. This resulted in it launching with the default URL, which
does not force experiments off. Then on the first load_tv() call, it
would switch to a URL that forced experiments off. The logic has been
changed so that an initial URL is provided and it is retrieved with
the same call as load_tv(), ensuring that experiments won't initially
be enabled and that they will stay in sync.

Change-Id: I969c9dca9b6d56f942ab5a707ab8aff8c6e7b6e8

Former-commit-id: f39017f5b6cebf4b850163f4525c7b13bc75151f