Various webdriver benchmarks improvements
The webdriver benchmark command line arguments have been modified to
disable the savegame (so that data from previous sessions won't be
used and data from this session won't be saved out), turn off the
debug console, and start Cobalt at "about:blank".

Additionally, the logic has been changed so that the first test now
runs setup logic, where the forced off experiments URL is loaded, and
then the reload that occurs afterwards is handled. After this, all
loads simply go to youtube.com/tv, since the "Forced Off All
Experiments" cookie has been retrieved.

To make sure that navigations are complete, the benchmark tests now
wait for the WindowDriver created notification before continuing with
any additional logic.

This should resolve the flakiness referred to in b/33275371.
Essentially, what was going on was that the Navigate does actually
wait for the Document to finish loading--the exact event that triggers
its return is Document::DispatchOnLoadEvent(). However, navigations
that newly set the forced off all experiments cookie would cause a
reload, which can cause the WindowDriver to become unavailable in
the middle of a test. This was inconsistent between machines and runs
because it depends on the state of the forced off all experiments
cookie. By disabling the savegame, we guarantee a consistent cookie
state at the start of the tests, and by waiting on the reload during
initialization of the tests, we guarantee that the reload won't cause
the WindowDriver to vanish during a test.

Change-Id: Ida948b69d69811ba4db59545dd87d1095f153c2f

Former-commit-id: 4c0c4c74517749094805989b41b93814645017e6