Add event tracking of generated computed styles
This was an oversight that this wasn't in before, largely due to the
UpdateComputedStyle() optimizations going in after the event tracking
being was added.

One of the main performance bottlenecks in UpdateComputedStyle() comes
from generating new computed styles. The optimizations that went in
last August (go/lbreview/55711) minimized the number of times that
computed styles need to be generated, so that it's no longer a 1:1
ratio between UpdateComputedStyle() and generating a style. As a
result of that, it's important to track the number of times computed
styles are being generated.

Change-Id: I92a1b727e81d1a5cfa2bbec0a3bf86da3a55b27b

Former-commit-id: 0851b010c732c47f2f1d3e15f5f96e569cd9a1fa