Fix event tracking with animation frame callbacks
Certain events with list 3.0 do not dirty the layout until after the
animation frame callbacks run (the main case for this is the first
horizontal navigation on a shelf). To handle this case, events are
now tracked if there are pending animation frame callbacks or if
there is a pending render tree. Additionally, there is now a callback
run when the animation frame callbacks are processed. If there is
still no render tree pending at that point, then the current event
being tracked is ended.

Change-Id: I0f185bf5155b6cf5b4c499e221792e8227fcb10c

Former-commit-id: a3c894331113c993289a6d609108a5c121bde14e