Improve webdriver benchmark URL loading logic
Previously, the webdriver benchmarks would wait on only the
WindowDriver created log. This works in most cases because the
SessionDriver does not return after a Navigation until the
WebModule has been loaded. However, when there is a reload, this does
not work because the SessionDriver has already returned. The order
that the WindowDriver is created and the WebModule is loaded is also
not deterministic, so waiting on only one does not guarantee that the
URL has fully loaded. As a result, the benchmark logic now waits on
both. An added benefit of this is that it is no longer reliant upon
the SessionDriver wait at all. Even if that logic changes, the
benchmarks will continue to work.

Additionally, improved the regular expressions being checked for in
tv_testcase_runner so that false positives can't happen (e.g. if a
video title contained the regular expression).

Change-Id: Ibda9a1a0d271c106ed5564435e54ec7593aac6fe

Former-commit-id: 0cd503c38add6ee7e30ceeb2c8ad98c4bf47260d