Parse and use index attribute in SkFontMgr_cobalt
Previously, the index attribute was not being parsed in
SkFontConfigParser_cobalt.cc and was assumed to be 0 in
SkFontMgr_cobalt.cc. This has been changed so that it is parsed and
used by the font manager, allowing Cobalt to support NotoSansCJK.ttc,
which is used on Android. However, we'll also want to add support for
font file streams, so that we avoid the hit from having
NotoSansCJK.ttc fully loaded in memory.

b/34106905.

Change-Id: I8076ab038c8c5b71da90dba3ec66d37c6e24843b

Former-commit-id: 020c4f7f891846d90b2b0bd33fee7964bff3d635