Fix memory leak in HtmlImageElement
Previously, if the src on an HtmlImageElement was updated while
an image was loading, the image loaded callback handler would get
destroyed, which prevented the associated callback from ever
occurring. However, garbage collection is prevented on the object
until the callback occurs, so effectively the HtmlImageElement
could never be garbage collected. This has been modified so that
AllowGarbageCollection() is now called if the src is changed while
an image is loading, fixing the memory leak.

This resolves b/35714205.

Change-Id: Ic78490b956682ed87f3b729a35a1ce1683f25274

Former-commit-id: 0a001972b95a2f8e1e4af30642c488b008403718