Send keys in benchmarks without specifying element
Previously, send_keys() in the benchmarks required an element to be
specified and sent the key through that element. However, this doesn't
match what Cobalt does when someone is pressing keys in the app. To
better mimic this behavior, the benchmarks have been changed to use
ActionChains, which allow us to call send_keys() without specifying
any element. Cobalt then will run through the same logic as it
does when someone presses a key while using the app, where it
dispatches it to the Window's active element.

Change-Id: Ieeb88dcb7787a0bd7d716252b3eb1e80b716ad2d

Former-commit-id: 83a2abc79ec7b9d6ffad4dc9a6f68891edfb9883