Clean up and rename time_to_shelf test
The time_to_shelf test has been renamed startup and the names have
been changed to better match the other webdriver benchmark tests.
Additionally, the blank to browse test has been changed so that it
tracks the 25th, 50th, 75th, 95th, and mean of the values and now
loads a blank page prior to loading the tv app each time, to
minimize time spent destroying the previous page.

Lastly, standardized the names across all of the tests, including
csi.

The db columns will be updated with the new names prior to this CL
being submitted.

Change-Id: I0a9c96f16c9483876913add53ce8dcb2b9459394

Former-commit-id: 1afba33efdd4675dc38fa2499710d93d728fce1e