Add a leak suppression for DirectFB.
While it appears to be the same leak that was supposed to be
already suppressed (261 bytes in 11 allocations), it looks like
on some buildbot slaves, the callstack comes out with a different
symbol, for some reason.  We add that here.

Also, the former library appears at the top of each callstack, so
it was actually suppressing all allocations.  This change makes the
original suppression more specific.

b/36001330

Change-Id: Id183e89e5f432ff91facfaba5a1489e4aa4af863

Former-commit-id: c686446aab0c5cf96b80cfcd4197483abdf1cd0d