Add Raspi app launcher
The Raspi app launcher is now fully functional. It uses the RASPI_ADDR
environment variable as the IP address for the test device and assumes
that the name is 'pi' and the password is 'raspberry'; they are not
currently configurable.

One thing to note is that PubkeyAuthentication currently is expected
to be turned off on the raspi device for the tests. This is because
the authentication triggers a request to touch the gnubby, which will
not work with automated tests and is a hassle with local testing.

Additionally, the pexpect is required and can be installed with the
following command: 'sudo apt-get install python-pexpect'

This resolves b/33112981.

Change-Id: Ic390559fede121ae7de9e0ed43c19596c19f8d8f

Former-commit-id: bf891aca339c394acd055902921c1c265a002604