Track max value of counts during benchmarks
Previously, the counts were only tracking the median and mean. This
has been changed so that they also track the max.

Additionally, Max and Min record strategies have been added and the
Percentiles one has been removed, since it is more straightforward
just manually adding the desired percentiles when constructing the
recorder.

Lastly, removed tracking of CntDomHtmlElementsDestroyed and
CntLayoutBoxesDestroyed. In the html elements case, the value is
pretty much always zero, as the tests don't run long enough for
garbage collection to be triggered. With the layout boxes, its easy
enough to get this number from other values that are available, so
this isn't useful to track.

Change-Id: I24c0b3fb4d0a38a4fe81aa04d31b036ab482ccae

Former-commit-id: 12aba517fab30da7047b86c8d5cad1ce899bac34