Websockets support
This CL adds websocket basic support in Cobalt.  More specifically, this
CL enables Cobalt to act as a client, and talk to another host using the
websocket protocol.  Compression, and other extensions are not supported
(on purpose).

This CL passes all of the normal tests from Autobahn's test suite.  It
doesn't pass some cases when client sends bad data (e.g. if the client
sets some reserve fields, etc.), and that can be easily enhanced in the
future.

b/27098590

Change-Id: I471db86783acb6acecc5120fbb82fdeb0be9eddc

Former-commit-id: 0019ee63fb94e1e3892e14ddb22efe3655259d57