Android: Handle accessibility changed
When Android's high contrast text mode is toggled, send a
kSbEventTypeAccessiblitySettingsChanged event.

Part of b/35669616

Note that while Android has a notification interface for when this
setting changes, it's hidden by an @hide. Since it requires a Java
interface definition, it cannot easily be accessed via reflection.

So we poll the state whenever our app gains focus.

Separately, but discovered at the same time, there's a race possibility
where the AndroidApplication might not yet be created, even after
"super.onCreate()" returns.

To resolve this, a mutex/condition was added to ensure that
NativeActivity.onCreate() does not return until after
AndroidApplication::Get() can successfully be invoked.

Change-Id: I775fbaa1d69dee1f5b1555f6f612a083db9dc39d

Former-commit-id: 68dd75e158908459fc842a61fb6009286bfa34f3