Fix transitions not stopping when the 'transition' property is removed.
When the 'transition' CSS property is removed from an element,
it will revert to the default of transition of 'all' but with
a duration of 0.  The original logic would only cancel a transition
if the property was explicitly not in the transition list, versus
being in the list (e.g. through 'all') but having its duration set
to 0.

Change-Id: Ie72f4253021ea412e84cdf02ddc759b5183ecfac

Former-commit-id: b2ff7ad7144a4b52414e165e156ca2e2ccab9973