Avoid potential shutdown crashes related to Camera3D/InputPoller.
Both the Camera3D object and the InputPoller might be attached to
a render tree, yet both of those objects will be destroyed before
the renderer is destroyed, resulting in the renderer holding on
to references longer than it should.  Fix both of these classes
against this problem.

The Camera3D class now forwards all calls to a Camera3DImpl object
which is *not* a script::Wrappable, and is therefore allowed to
outlive the WebModule object that created it.

b/36079419

Change-Id: Ida47ef1bafdaddefc486a95152a76f3516ac7449

Former-commit-id: b7e72f71d015532fa8ab0f653512d4970af38c8a