IDL dictionaries: sequence support
...at least inside an interface...

Previously: A sequence<DictionaryType> would fail at compile time
because the From/ToJSValue's for the dictionaries
weren't declared before the the From/ToJSValues for the sequence.

Now:

Just declare the dictionary From/ToJSValue'a before the conversion_helpers.h
where the sequence From/ToJSValue templates are included.

Note:

This does not seem to work for dictionaries that contain
sequences of dictionaries.

Also note:

Similar work needs to be done for mozjs-45

Change-Id: I1f504f3b400754de2013b148c329da92e4e2c074

Former-commit-id: 872a4f86cb6c7494b52be70cf0a9d65dadebda94