Fix renderer test failure.
We would like to clear the screen to black before quitting Cobalt
in case some platform does something funny with the leftover
framebuffer.

However, this resulted in test failures that expected the last
rendered frames to remain.  This CL makes it so that the screen
clear is an optional pipeline parameter.

Change-Id: I98d73f422f81cfe0bffa021440f516e6851c8392

Former-commit-id: afc8e5f215cd5e023bcea4639589fc73ac9fcd3c