Make some fonts.xml attributes optional
To make fonts.xml more similar to the android version and easier to
use, some of the attributes have been made optional.

With family, the fallback attribute is now optional. If it does not
exist, then Cobalt follows the same logic as Android where the lack
of a name dictates whether or not it is a fallback family. The
fallback attribute allows the user to override that functionality and
make a named family a fallback.

With font, font_name and postscript_name are now optional. They don't
provide any utility with the fallback fonts and removing cleans up the
xml and shrinks the size of the hashes in the manager. If these are
not provided for a font, then it cannot be matched by face name.

Change-Id: I7a0c94122124627eb3c377f028a8ba4427ab388c

Former-commit-id: a4bd830f31397acdb09bc11ff0b7f899256f4269