Remove CarroisGothicSC from smaller font packages
CarroisGothicSC was previously in the minimal and 10mb font packages
to satisfy the FCC requirements, since ***REMOVED*** wasn't serving a
smallcaps font that was usable by us. They've since fixed this and
send down the same file that we package, so it is no longer needed
in the packages.

The ***REMOVED*** CL fixing this is:
https://critique.corp.google.com/#review/136391185.

Change-Id: I4cf44f02e9959ef6665da5b03c2e90d5ab06d7f9

Former-commit-id: 57245b297569f19e63f2ff99d8c3fec9d31a06d5