Use an atlas for offscreen targets with direct-to-GLES rasterizer.
Start with a small (relative to the render target) texture atlas and
grow it until it can handle all offscreen targets requested in any
given frame.

The DrawObject::ExecuteRasterizeOffscreen has been renamed to
ExecutePreVertexBuffer and moved to process before
ExecuteUpdateVertexBuffer so that texcoords can be adjusted to
coincide with offscreen targets in the atlas.

Change-Id: I83e9a65015791093da2b5e7f43fdf2c0e59e4230

Former-commit-id: bc90145d4008c1c4c778ef06c38ae63b694c0033