Fix and rename MinimalRoboto and improve generation script
Previously, MinimalRoboto was not including the correct tofu glyph
for Roboto and it was including a large space after the tofu. This
would cause issues when ***REMOVED***'s Roboto loaded, because the spacing
of the default font and the web font would not be similar, which
would cause a large shift in the layout. This has been fixed so
that the file uses the proper glyph.

Additionally it has been renamed Roboto-Regular-Subsetted, so that it
follows the same naming conventions as the rest of the fonts.

The script for generating Roboto-Regular-Subsetted has been modified
so that it no longer goes through glyphIgo and instead directly uses
FontForge, which significantly simplifies things and enables us to
remove all of the later steps involving adding.notdef to the generated
font.

Change-Id: I3532391b7c8b2fc1f29d26fbc449c5c096e10b4e

Former-commit-id: a7d4b9c994079dbb6d083634ee36cc4be09c7b99