Fixes memory tracker test on linux.
Older versions of clang will actually eliminate operator new.
This causes problems since operator new will generate an event
for memory allocation which are necessary for the tests.

This is a bug because the standard says that functions should
not be removed if they have side effects. This fix uses volatile
to break aggressive optimization.

Change-Id: I4f14098085564bb145bc0d27b737a2b88484c87e

Former-commit-id: b5fc6284f958448a963521f9e37d043a145848b6