Fix renderer_test DCHECK failure for direct-to-GLES rasterizer.
This restores the OpenGL graphics state back to defaults when outside
of a render frame. This fixes the renderer_test failure in
GraphicsContextEGL::ComputeReadPixelsNeedVerticalFlip which needed
depth testing disabled to work.

This also makes it so that GraphicsState::Reset() actually restore the
currently-cached state rather than resetting it to some default. This
makes SetDirty() work properly if called while rendering onscreen.

Change-Id: Ia607bda872f0e36d60ed46d881228ddb0738f38a

Former-commit-id: 77babc3616335653413d9f7dbb1b58507e15cafb