Remove purging of images when invalidating layout boxes
Previously, HTMLElement::InvalidateLayoutBoxes* was calling
ReleaseImagesAndInvalidateComputedStyleIfNecessary(). However, this
clears the background images on an HtmlElement and invalidates its
computed style. Given that box invalidation is often done after the
computed style is calculated; this is not ideal. It resulted in bugs
where the background image on an element would be set and then
immediately cleared. The image would then not appear until another
UpdateComputedStyle round occurred.

Invalidating layout boxes no longer clears the cached background
image.

Instead, clearing it only occurs in three cases:
1. An HtmlElement has its display change to None. In this case, all
   of its descendants have their images purged, because they are no
   longer appearing in the active DOM tree.
2. The app is suspended. In this case, Document::PurgeCachedResources()
   handles clearing the images.
3. A node is removed from the document. In this case, we want to make
   sure that we aren't holding on to references to images that can now
   be released. This is now being handled by an explicit call to
   PurgeCachedBackgroundImagesOfNodeAndDescendants() in Node::Remove.

Additionally, a fair amount of cleanup has been done with the names of
the various recursive calls in Node and HtmlElement.

This also fixes a crash that was our fourth most common crash in
production, which was related to the invalidation of layout boxes
triggering DOM events, which triggered the invalidation of layout
box state on deleted boxes.

This resolves b/36540325.

Change-Id: I8ac0d758e5ec2ce69503f1b4b1ccf58211b25ac5

Former-commit-id: 2923fb7226e01889c8f1bd0d28a37ddabddc078a