Do not purge image cache when lowering capacity for video playback.
Purging the entire image cache when lowering the capacity will
result in a complete reload of image data upon returning to
the UI from playing a video, which is an unpleasant experience.

Note that this feature is currently not used on any platform, but
it may be desireable to save memory, and purging the entire image
cache makes it a difficult option to consider.

Change-Id: I3f49cd67b694b3c35fafa172da67a8735dad736d

Former-commit-id: 27f1d063c47edd1254a218168a896361209292e9