Use PS3 JPEG decoder if required #define exists
The PS3 JPEG decoder will now be used if USE_PS3_JPEG_IMAGE_DECODER
is defined. It was previously using a define that stopped existing
after the switch to Starboard, which broke the PS3 JPEG decoder.

This is not the long-term solution. That is to switch the decoder to
going through Starboard. The bug tracking the long-term fix is
b/33753933.

Change-Id: I56e22a8ef88f4c7c3e1914eae9c0408f2eb6a078

Former-commit-id: 21152afe18af9061f81eca30e4ae0405c220b707