Start Loaders in a suspended state when factory is suspended
Previously, when Suspend was called on a loader factory, it would
suspend current loaders, but any new loaders that created while
suspended would not start in a suspended state. These loaders did not
have resource providers, but would still download all of the resources
and do whatever decoding work was allowed without a resource provider.

This logic has been changed so that all loaders created while the
loader factory is suspended are created in a suspended state.

Change-Id: Ic6acc767755b6c4eb3a5ccd07515855a8136d4f2

Former-commit-id: 6a79b684259e4024c1a72094d7b77e1054f9069d