Add Android specific video renderer
This change adds an Android specific video renderer that is capable
of properly handling the releasing of output buffers and reporting
state to the player worker handler.  This fixes the video freezing
issue that was occuring on Nexus Player, since the platform's media
codec implementation did not have enough output buffers to preroll
to the extent that the old shared video renderer wanted to.

As a result of this, the |Host| component of the video decoder
interface has been pushed down into a new class, |HostedVideoDecoder|.
The ffmpeg and openmax decoders now implement this interface instead,
and the old shared video renderer now works with this interface.

Logic has also been added to handle the case in which we attempt to
decode input buffers with DRM info before we have resolved our keys.
The keys were getting resolved so fast when developing on the Shield
that this case was not previously getting hit.

b/36568108
b/36280721
b/35708612
b/34822842
b/34354465

Change-Id: I9472061be71b3f71483394c34ff18e4d87521cd7

Former-commit-id: 25224434319b9f410ef540049edc04c6a1e8e8ae