Fix WebModule resource cache logic to include MeshCache
Previously, the resource cache purges that happen on a suspend and
the disabling of callbacks that happen on shutdown were not being
called on the mesh cache. The mesh cache has been added to both cases
and the purge and disable logic has been moved into its own function
to make it more obvious that it should happen on all resource caches.

Change-Id: I312148296183c1c969f8b727109ac7f4e70699a5

Former-commit-id: 3f6cb66489bfcbec6628db84ebb55d067228901e