Allow dirty region rendering to be toggled in gyp_configuration.gypi.
On some platforms, calling
eglSurfaceAttrib(..., EGL_SWAP_BEHAVIOR, EGL_BUFFER_PRESERVED) can
result in an extra full-screen surface allocation being made on the
GPU, and this may not always be worth the performance improvement
that it brings.

This CL adds a flag so that platforms can optionally disable dirty
region only rendering.

b/36577982

Change-Id: Icd7376617687e5bb24496f047beb7a06f1ea5e61

Former-commit-id: bcd73b39853ff0d3c8cf7459d4eb61744d41048f