Improve Android video playback
This change improves the framerate and playback experience on Android
devices by modifying the order that work is done in the video
decoder.  Rather than switching off between doing as much media codec
work as possible and then handling frames that are ready, we instead
attempt to handle output frames for each media codec buffer (both
input and output) that we work with.

Change-Id: Id26ee57ccf3f553312e56d61d6e06c35d1a0d1e6

Former-commit-id: 00b4f7a03a912d58b207b2d09f383677564989b9