Enable websockets implementation in Cobalt
This CL enables WebSockets implementation in Cobalt.  Before this CL,
the WebSOcket::Connect() was not calling the WebSocketImpl.  This is
because the Impl would always create real sockets.  This prevented the
tests WebSockets from running if the WebSocketImpl::Connect() was
uncommented.

Now, the WebSocketImpl can handle a NULL network_settings, so the test
code can pass by just passing in NULL for the network_settings.

Change-Id: I07cda3418084903faf33a51f8e1a6b8303a1286b

Former-commit-id: d03131c6f6aac55c3b610cf6b16f6b9f590d0dd4