Use the system allocator instead of dlmalloc for Linux.
Unfortunately our EGL libraries seem to have a bug in them where
they try to free something through dlmalloc that was not a valid
dlmalloc allocation.  Switching to the system allocator instead
of dlmalloc fixes this problem (presumably because the thing that
was attempted to be freed was somehow allocated by the system
allocator).

Also, it has been found that on some Linux-based platforms,
the system allocator can perform better than dlmalloc.

Also disable ASAN so as the EGL libraries also seem to introduce
leaks.

b/36807034

Change-Id: I4863aa4bddd8362ca838b4660d44a328103a9f91

Former-commit-id: 4ebce1e1010f35dbc08aa3e69cf75101060bc776