Ensure that a depth buffer is present for the direct-to-GLES rasterizer.
This change is necessary with the new FramebufferRenderTargetEGL.
The graphics system's config would make all PBuffer render targets have
a depth buffer as needed, and this logic could be carried over to the
FramebufferRenderTargetEGL. However, that approach felt more restrictive.
Instead, make the rasterizer enforce creation of a depth buffer if it
needs one.

Change-Id: I3b27a68fa8805caf951a572f6b4fb2f858e2f819

Former-commit-id: 067483b8ad5589ac767f0d0ae909915172ddedac