Fix opacity handling with direct-to-GLES rasterizer.
This addresses a subtle misunderstanding between rendering objects with
transparency vs. rendering an offscreen target with transparency. The
former was being done, but this produces incorrect results when objects
overlap within an opacity filter. The overlapping objects should occlude
each other as appropriate, then the entire result rendered with
transparency.

Change-Id: Iefc43bea326920156a144308fd11c128c3cafdec

Former-commit-id: 31d28e9a697409329645f09390f42b62caeec9f5