Check if video is 360 closer to when the player is created.
While the original solution turned out to be fine, and there was
an issue on the player side, it is safer to give JavaScript
as long as possible to attach a map-to-mesh filter to the video
element.  So, we wait until the first data is appended to
media source and then do one last update of if decode-to-texture
is preferred or not.

Change-Id: I8dbe44746ad64a25b83285fab3a0b54402786573

Former-commit-id: 0e0cd832559c54c6d9d9b5e0bad08cff83dfce50