Tweak handling of the offscreen render target for fallback rendering.
If the desired offscreen render target would be larger than the main
viewport then clamp the offscreen target to the size it would occupy on
the main render target, and only render what will appear in the main
render target. This helps avoid pixelation.

Also fix the offscreen render target manager so it will never try to
allocate a rect with 0 dimension(s).

Change-Id: Ifdd8cdb446fb440dc10d584120928e7c2d8ada85

Former-commit-id: 03f21f5cb18ff596f4dacf4dd580c49898a4c3d2