Fix live videos
GetElementsByTagName was recently changed to do a lowercase search for
the element. However, this is only valid for HTML documents and not
for XML documents. As a result of this, XML documents were unable to
successfully look up elements by the tag name, which resulted in all
live videos being broken.

The logic has been changed so that the conversion is only done for
HTML documents.

Change-Id: Ida29eee949cea990b7a6509b1a420f6f98f76363

Former-commit-id: f884fe68868a453f30fe4af1a6b0ee00e9293e87