Fix Android input latency during animation playback.
By default, Android sets up a graphics buffer queue that
Cobalt produces into (when eglSwapBuffers()) is called and Android's
SurfaceFlinger consumes from (each vsync).  Unfortunately, this queue
seems to be large enough (empirically it looks like ~3 frames) such
that it can affect input latency.  This CL calls eglSwapInterval(0)
to put Android's SurfaceFlinger into "async mode" where instead of
using the queue it just updates the "current" buffer with the result
of eglSwapBuffers() each time it is called.

Additionally, we set the minimum frame time on Android (and all
devices) to 16.4ms, slightly under 60Hz (16.666ms), so that we can
ensure that we are always rendering at least as fast as 60Hz, if
we are able to.  Since we are managing the frame interval ourselves,
we call eglSwapInterval(0) on all platforms now in case other platforms
do something similar to Android (and for those platforms that don't,
this will not harm anything).

This CL also replaces the frame rate throttler with simply updating
the renderer's timer.  This reduces a lot of overhead code when
using the framerate throttler and also frees up the thread to handle
other tasks (like texture updates) when it's not rendering a frame.
Doing it this way also means that we don't delay before calling
eglSwapBuffers(), so the produced frame is available to display
sooner.

b/37157623

Change-Id: I467b1e96ed858746bfd5984bbb4b70264d061677

Former-commit-id: 9b33518b9cfeebb3a72cddeb06c0ca96f7dfb3ff