Fix heap use after free
Previously, there was a race condition in the websockets code in Cobalt.
This CL fixes it.

Race condition: If a message was sent from a network thread, and there
were no javascript event handlers registered, the websocket object would
be destroyed, and the callback would read freed memory.

Now: The main websocket object is still destroyed, but calling back into
it is prevented because callbacks from WebSocketImpl are now disabled.

Fixes b/36865033.

Change-Id: I5ab6e6e29ab5cb5ce63fc998c687c894e8c83915

Former-commit-id: 9d5e8e90e2e6601c9f13df7573c1b849233d217c