Move ScreenReader to H5vccAccessibility
We need to disable ScreenReader at runtime.
The obvious way to do that is via an H5vcc attribute.

Unfortunately, having H5vcc depend on WebModule is not feasible.

So, instead, move ScreenReader into H5vccAccessibility.

An enable/disable attribute will be added in a followup cl.

This will be cherry-picked into COBALT_9.

Change-Id: Ia97a856f3da637d4961d4d7577f7f7bb7ba884fa

Former-commit-id: 27fb3c793ad8a36dc040b6d8ba000e11fcda9902