Add RefCountedLock and resolve CVal lifetime bugs
RefCountedLock and RefCountedThreadSafeLock have been added to base.
They are simply refcounted wrappers for locks.

Modified the value lock in CVal to use RefCountedThreadSafeLock, so
the various CVals can continue to use the value lock even after the
CValManager has been destroyed without requiring each CVal to create
its own lock.

This allows CVal to gracefully handle the case where the CValManager
is destroyed prior to all of the CVals. Previously, this would cause
a crash. This was an issue that prevented CVals from being owned by
SkFontMgr_cobalt, because it can have a longer lifetime then the
CValManager.

Change-Id: I7975d6b31be1bd1ba9d057e13502bb72dc343936

Former-commit-id: ef7efe6893b6f8cc86abd0023f0d591a5b3f46a7