Replace .otf CJK files with NotoSansCJK-Regular.ttc
With the switch over to SkFileMemoryChunkStream, Cobalt can now
support using NotoSansCJK-Regular.ttc. This allows us to provide the
same fonts as Android, will enable us to easily offer a bold version
of the CJK fonts, and offers more consistent fallback choices with
CJK languages. The one downside is that it adds 2MB to the package
size.

Change-Id: Ic720c7e3e9167205e34c4f52f6e28564e9257410

Former-commit-id: 2bd4c029f0ba42bb58e6f20e8c59d71de25aca23