Fix bug in SkOSFile_cobalt
Previously, sk_fopen in SkOSFile_cobalt was not checking the
base::CreatePlatformFile return value and was instead simply calling
reinterpret_cast on it to cast it into a SkFILE pointer. While this
works properly for the value that OS_STARBOARD uses for
kInvalidPlatformFileValue (NULL), it does not work for either OS_WIN
or OS_POSIX, which use a value of -1.

By checking the value against kInvalidPlatformFileValue and always
returning NULL if it matches, then we're guaranteed to have the same
functionality across platforms.

Change-Id: I691528ed0a5489e2f7e15542ba50e9c999b4f07c

Former-commit-id: fc9377a4b21a441711a61732f0648f5384be762e