Add aux field to user agent
On Android, we need to report the app package and version name.

We're going to do that by appending it to the user agent.

We've added kSbSystemPropertyUserAgentAuxField to do this.

This adds the aux field to the end of the user agent, if
available.

The result is a user agent (for coat on Nexus Player) like:

Mozilla/5.0 (Android; Linux x86) Cobalt/10.42065-debug (unlike Gecko)
Starboard/5, _ATV_moorefield/LMY47V (Asus, Nexus Player, Wireless)
foo.cobalt.coat/1

b/34613667

Change-Id: Id68ba07baaafd5e858d77af5b583028a95dec099

Former-commit-id: 18f8d5b1afa530e1598a8464ffcc1b56ae24d7c3