Remove incorrect tests for IFA_F flags.
This removes the tests against IFA_F flag values (values for
temporary and deprecated) for the ifa_flags member of the interfaces
returned by getifaddrs(), because they are not valid for this flag
field.  IFA_F_TEMPORARY (in iff_addr.h) has the same value as IFF_UP
(in iff.h).

Flags values returned by getifaddrs are those from SIOCGIFFLAGS,
documented in 'man netdevice'.  IFA_F_* flags are for linux netlink
(https://tools.ietf.org/html/rfc3549), which is a non-posix, and
linux specific API.  Since implementing this is nontrivial, filed
b/37422542 for addressing it.

Change-Id: I1918ae870f44994ce595a139f023860e0b395f5b

Former-commit-id: a701ccdc394dab932e96e624dc102bdf87da3b43