Reduce max memory usage of the direct-GLES rasterizer.
This adds a memory limit parameter to the offscreen target manager. In
order to handle hitting the memory limit, render target allocation has
to be done by the render tree node visitor so it can change when the
render target should be rendered to (during the "offscreen" phase or
just before it is used in the "onscreen" phase).

This also moves the transform into the fallback rasterizer to improve
sub-pixel correctness.

Change-Id: Ia3e9d7927d29bb6e79764f2279d3e6aacfc676ed

Former-commit-id: 22b215f128530d53037d69041e34d8c91af9987d