Add support for synthetic bolding
Cobalt now supports synthetically bolding non-bold fonts when the
desired weight is greater than 500. However, there is also a new
attribute in fonts.xml "disable_synthetic_bolding" that provides the
ability to selectively disable this functionality. This disabling is
currently used with DroidSansFallback as a result of its Traditional
Chinese glyphs blurring together when synthetically bolded.

Layout tests have been added for this.

Change-Id: I49d621c7e76094cf6dcfe042a0f01537678a1fd0

Former-commit-id: 5a0e7fbf2368dab5376fb5c8edf377d02734b85e