Font package cleanup
Updated the Cobalt font packages to include 5 variations, rather than
3.

Packages:
1. 'expanded' - This is similar to the old 'unlimited' package, but
   includes bold Noto CJK. This is ~46.2MB.
2. 'standard' - This matches the old 'unlimited' package. It is
   ~26.4MB.
3. 'limited_with_noto_jp' - This matches the old '10megabytes'
   package. It is ~10.9MB.
4. 'limited' - This mostly matches the old '10megabytes' package but
   removes the Noto JP file, and relies exclusively on
   DroidSansFallback for CJK. It is ~7.6MB.
5. 'minimal' - This matches the old 'minimal' package.

Modified fonts.xml so that it is shared between 'expanded', 'standard',
'limited_with_noto_jp' and 'limited' so that multiple fonts.xml files
don't need to be kept in sync with each font change. Fonts that are
unavailable within a specific package will simply be stripped out on
startup.

Cleaned up copy_font_data.gypi so that the various font files are
divided into groups, which makes it easier to follow what is in each
package. Also, added mappings of the old package names to the new
names for 'unlimited' and '10megabytes'.

Change-Id: Ib0554197ce7413b8badecc0d2df1efe9a60336ab

Former-commit-id: f5273fecd243b0b003b78f77952233d0b67ad6fd