Android: Support DPad when TalkBack is on
When the Android TalkBack screen reader is on, it intercepts
DPad events, causing the user to be unable to interact with our
application.

In lieu of annotating our render tree and exposing it to the
Android accessibility framework (which would be significant effort),
we simply expose a "nine-patch" of regions to the accessibility
framework for TalkBack to navigate, translating the changes in focus
here to dpad events in Cobalt. Cobalt then reads out the appropriate
accessibility text.

Existing bug with this checkin: the first dpad event is ignored, as
focus changes from the "host view" to the nine-patch.

b/35946320

Change-Id: I8e7a3856b957adff5e8aa5f84be40d28caceebc7

Former-commit-id: b6256513559ea45777531aa3f5328a44ea4a0c60