Give porters the ability to customize font packages
The font package logic has been modified so that the basic package
values can now be overridden if porters want to modify them.

This provides porters with a simple path where they just use one of
the packages that we provide, or a more complex path if needed,
where they use one of our packages and then override some of the
categories to exactly fit their needs.

The current categories are as follows:
  'package_named_sans_serif'
  'package_named_serif'
  'package_named_fcc_fonts'
  'package_fallback_lang_non_cjk'
  'package_fallback_lang_cjk'
  'package_fallback_lang_cjk_low_quality'
  'package_fallback_lang_jp'
  'package_fallback_emoji'
  'package_fallback_symbols'

Each category can have a value from 0 to 4: They indicate the
following things:
  '0' -- No fonts from the specified category are included.
  '1' -- Fonts from the specified category with a weight of 'normal'
         and a style of 'normal' are included.
  '2' -- Fonts from the the specified category with a weight of either
         'normal' or 'bold' and a style of 'normal' are included.
  '3' -- Fonts from the the specified category with a weight of either
         'normal' or 'bold' and a style of either 'normal' or 'italic'
         are included.
  '4' -- All available fonts from the specified category are included.
         This may include additional weights beyond 'normal' and
         'bold'.

Change-Id: I4d94e7edbd26b4c1beb7ee0b5c0c00c56e083625

Former-commit-id: 72f881588905b4d69d42fe9c870e2a0117d1818c