Add support for system fonts
In addition to its packaged fonts, Cobalt now has the ability to load
system fonts from directories specified in Starboard. The only
requirement with it is that font names from fonts within the package
cannot be duplicated in system fonts. However, given that porters have
the ability to strip out selected portions of Cobalt's font package,
this should not be an issue.

Additionally, fallback fonts now have an integer fallback priority.
This dictates the order that they are considered during fallback and
allows system fonts to be selected first during fallback if they are
given a priority greater than 0.

Lastly, stripped out most of the aliases that Cobalt was adding to
fonts. This was mimicking Android's strategy, but it didn't really
serve a purpose for us. The last family in font-family lists is
supposed to be a generic family, and we still include all of those,
as well as all of the FCC font names. So we should still successfully
match every font-family, just possibly with a later family in the
list. The benefit of doing this is that there are fewer potential
name conflicts with system fonts of porters. We're now providing the
core families and nothing more.

Change-Id: Ia72342372b3b244f9a696b503db8293a55f5e311

Former-commit-id: f117178e79759b10c7bf8e5f738cf0db98be242b