Add call to SbSpeechSynthesisCancel.
This adds a call to SbSpeechSynthesisCancel in
SpeechSynthesis::Cancel() to stop the current utterance from being
uttered any further, if there is any.

Since there is no knowledge whether an utterance is actually being
spoken at the time of the cancel, there will never be a "canceled"
speech synthesis error event for the interrupted utterance.

Change-Id: I744caba351877e1d4fe91f4bdadefbbbcda693f7

Former-commit-id: 6b2acc57ca96127d3efb0f2ba8e8d689e82e3063