Fix Wrappable/Wrapper garbage collection issue
Previously, |Wrapper|s of |Wrappable|s were not referenced anywhere,
which resulted in the SpiderMonkey garbage collector frequently
garbage collecting them. In some cases in which the wrapper was
needed again by JavaScript, this was ok, since a new |Wrapper| could
be constructed for the native |Wrappable|. However, if any custom
properties were added to the |Wrapper|, they would be gone on the
newly constructed |Wrapper|.

This change adds logic to properly mark |Wrapper|s that |Wrappable|s
refer to (by referring to other |Wrappable|s) as reachable, via a new
|TraceMembers| function that is part of the |Wrappable| interface,
and a |cobalt::script::Tracer| class that performs the necessary
traversal.

b/37744136
b/37731263
b/37571108

Change-Id: I712eea41d1e7f16168765b587d60ac4bfec45ddc

Former-commit-id: 269a9f3c88cccfaf401897c3806432738a685f65