Fully purge font caches during Suspend()
Previously, the font stream data and Skia glyph cache data was being
purged on suspend, but the remainder of the font caching was retained.
While the remaining caches are not as large, they can still easily
account for several hundred KB in memory (and potentially MB worth).

Now, all font caches are purged. This includes the Dom's FontCache,
which will allow the remote typefaces to be purged, Harfbuzz's
FaceCache, and the unreferenced typefaces within SkFontMgr_cobalt.

The purging of the resource provider's caches has also been moved so
that it occurs within its destructor. This allows it to occur after
all render trees are destroyed, which ensures that no typefaces are
still being referenced by glyph buffers.

Change-Id: I09445c788dd35fb71d20fba6308ae7123952228c

Former-commit-id: 6631e2e9db80e7c63ceef30395c847482973d768