Hookup mouse events from X11 to Starboard.
This enables mouse button and motion events for input to the X11
application and converts them to the corresponding Starboard Input
Events.

For platforms where the Starboard API is older than
SB_POINTER_INPUT_API_VERSION, wheel events are reported as key
presses, and chorded button interactions are not reported as such due
to limitations in the API.  See go/cobalt-mouse-support for details.

b/37275032

Change-Id: Ie0cc0d901b138f4316a59108e669d657339ad7cf

Former-commit-id: cb6ac4be38cfb07635a6f738149e09ade082098f