Add a serial number to render targets.
This CL adds a serial number to each render target that is created.
The serial number acts as a good key in hashmaps.  Previously, the
rasterizer was using the memory location of a render target as the key,
which could have a collision if a render target was re-created at the
same address.

Fixes b/37569671.

Change-Id: Ia17080fb179218f82fc001d9bad41f07e335990c

Former-commit-id: 0d849f5b30bda4884611d608e114719080b6afbe