Add a serial number to render targets.
This CL adds a serial number to each render target that is created.
The serial number acts as a good key in hashmaps.  Previously, the
rasterizer was using the memory location of a render target as the key,
which could have a collision if a render target was re-created at the
same address.

Fixes b/37569671.

Change-Id: Ia17080fb179218f82fc001d9bad41f07e335990c

Former-commit-id: caa1884a650de1f0bf68d8480f9dffb3222044f2