Add walk_dir.py
It contains a framework that iterate through all files in a given
folder, check each file names against a filter function and run a
processor function on the file names that pass the check.
It comes with pre-built filer functions to filter C++ source code
files and general source files.  It also comes with a few processor
function, like the one that can replace license header or starboardize
C functions.

Change-Id: I92ddea9c6202422013251ea9a51291685f2ec814

Former-commit-id: e3817f83722f2c745ab42b1205d831155c85e238