Fix progressive video seeking.
ShellDemuxer::SeekTask() used to call Request() before cleaning up some
of the variable indicating eos status.  The original Request() posted
a RequestTask() so it always run after SeekTask() is finished.  A later
refactor of Request() won't post a task if it is already on
|blocking_thread_| and will run before SeekTask() is finished, which
causes seeking issues on progressive videos.

b/38502493

Change-Id: I181bce274b606febb184d8fbd3d9c2358856b8a4

Former-commit-id: fc1bb51498c2b6b7bd77ba9876c358f0f32cbc23