Remove DCHECK in dom::Event::InitEvent().
InitEvent() should be called before the event is dispatched and any
further call to it after dispatching should be ignored.
The previous code also DCHECK on redundant InitEvent() calls and
triggers crash in ClusterFuzz.  It is safe to remove the DCHECKs as the
error conditions are also properly handled by condition statement.

b/38310625

Change-Id: Id6475aa581aa104feb66c4da49b13844ae3f19b7

Former-commit-id: cb1d81c6792d295c7dfa5edac3b5d764933a1e9e