Replace SB_DCHECK and SB_DLOG with DCHECK and DLOG
All uses of SB_DCHECK and SB_DLOG in MicrophoneFake are being replaced
with DCHECK and DLOG. The reason for this is that SB_DCHECK does not
work with base::ThreadChecker. It causes a QA build error for an
unused private field. DCHECK doesn't trigger this error; for
consistency, all SB_DCHECKs and SB_DLOGs in the file are being
converted to DCHECK and DLOG.

Change-Id: I68cdc19500389490a79fa5c68687a4119bfc86d1

Former-commit-id: 83b784204b0b46e24feabe24d2db6bf48962c99e